% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allCorners.R
\name{allCorners}
\alias{allCorners}
\title{allCorners}
\usage{
allCorners(cornerSet, outcomeLabel = "OUTCOME")
}
\arguments{
\item{cornerSet}{is an dataframe produced by prepNettle()}

\item{outcomeLabel}{is a passthrough variable for the name of the outcome}
}
\description{
allCorners allows you to plot all of the plots in a prepCorners() list
}
