% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawCorner.R
\name{drawCorner}
\alias{drawCorner}
\title{drawCorner}
\usage{
drawCorner(sector, nElements, drawTop, drawRight, leafColor, offset = 0.028,
  rElement = 0.02, unique = FALSE)
}
\arguments{
\item{sector}{is a predefined viewport}

\item{nElements}{is the number of elements to plot}

\item{drawTop}{is half of the definition of the corner of the viewport to draw the elements in (1 = top)}

\item{drawRight}{is the other half of the definition of the corner (1 = right)}

\item{leafColor}{is the color for that particular leaf}

\item{rElement}{is the size of the element}

\item{unique}{draws with white fill so that a fifth call can draw unique elements}

\item{offSet}{is the distance between elements}
}
\description{
drawCorner is a function that is used by fourCorners to draw one of the leaves.
}
