% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPathRecovery.R
\name{getPathRecovery}
\alias{getPathRecovery}
\title{getPathRecovery}
\usage{
getPathRecovery(QCAoutput, simOutput)
}
\arguments{
\item{QCAoutput}{the output from a QCApro analysis}

\item{simOutput}{the output from simCases}
}
\description{
What percentage of paths were recovered in the sample?
}
