% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simFuzz.R
\name{simFuzz}
\alias{simFuzz}
\title{simFuzz}
\usage{
simFuzz(numVars = 2, numCases = 50)
}
\arguments{
\item{numVars}{creates n vars that are all in fuzzy subset relationships. Must be greater than 2. Default is 2.}

\item{numCases}{is the number of cases to include in the generated fake data set. Must be greater than 1. Default is 50.}
}
\value{
A data frame with one column per variable.
}
\description{
simFuzz is a function to generate simulated continous data for cluster analysis
}
