% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourCorners.R
\name{fourCorners}
\alias{fourCorners}
\title{fourCorners}
\usage{
fourCorners(cornerSet, outcomeLabel = "OUTCOME")
}
\arguments{
\item{cornerSet}{is an dataframe produced by prepNettle()}

\item{outcomeLabel}{is a passthrough variable for the name of the outcome}
}
\description{
fourCorners allows you to plot all of the plots in a prepCorners() list
}
