% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCorners.R
\name{plotCorners}
\alias{plotCorners}
\title{plotCorners}
\usage{
plotCorners(pathOutcome, othersOutcome, pathNotOutcome, othersNotOutcome,
  uniquePathOutcome, path = "PATH", outcome = "OUTCOME",
  textPosition = 0.25)
}
\arguments{
\item{pathOutcome}{is the number of cases that have the path leading to the outcome}

\item{pathNotOutcome}{is the number of cases that have the path, but not the outcome -- used to calculate consistency}

\item{othersNotOutcome}{is the number of cases that are irrelevant (no path or outcome)}

\item{uniquePathOutcome}{is the number of cases that only have that path to the outcome (and not some other)}

\item{path}{is the name of the path}

\item{outcome}{is the name of the outcome}

\item{otherOutcome}{is the number of cases without the path but with the outcome -- used to calculate coverage}
}
\description{
plotCorners is a function that draws the full four leaf fourCorners plot
}
